Refactor SkillsFilter component to use custom hook
