+++ src/pages/EmployeeListing/components/SkillsFilter.tsx
@@ -1,8 +1,7 @@
 import { useEffect, useState } from "react";
 import { Skill, Employee } from "../../../models";
 import ClickAwayListener from "../../../components/eventListeners/ClickAwayListener";
-
-type SkillOption = Skill & { count: number; checked: boolean };
+import useSkillsFilter from "../hooks/useSkillsFilter";
 
 export interface SkillsFilterProps {
   skills: Skill[];
@@ -17,39 +16,14 @@ export default function SkillsFilter({
   selectedSkills,
   onChange,
 }: SkillsFilterProps) {
-  const [isActive, setIsActive] = useState(false);
-  const [options, setOptions] = useState<SkillOption[]>([]);
-  const [searchTerm, setSearchTerm] = useState("");
-
-  function onClickSkillOption(skillId: string) {
-    if (selectedSkills.includes(skillId)) {
-      onChange?.(selectedSkills.filter((id) => id !== skillId));
-    } else {
-      onChange?.([...selectedSkills, skillId]);
-    }
-  }
-
-  useEffect(() => {
-    const options = skills
-      .filter((skill) => {
-        return skill.skill.toLowerCase().includes(searchTerm);
-      })
-      .map((skill) => {
-        return {
-          skillId: skill.skillId,
-          skill: skill.skill,
-          count: employees.filter((employee) => {
-            return employee.skills.find((skillItem) => {
-              return skillItem.skillId === skill.skillId;
-            });
-          }).length,
-          checked: selectedSkills.includes(skill.skillId),
-        };
-      })
-      .filter((option) => option.count > 0)
-      .sort((a, b) => b.count - a.count);
-    setOptions(options);
-  }, [employees, skills, selectedSkills, searchTerm]);
+  const {
+    isActive,
+    setIsActive,
+    options,
+    onClickSkillOption,
+    setSearchTerm,
+    onClickClearFilters,
+  } = useSkillsFilter({ skills, employees, selectedSkills, onChange });
 
   return (
     <ClickAwayListener onClickOutside={() => setIsActive(false)}>
@@ -125,7 +99,7 @@ export default function SkillsFilter({
             })}
           </ul>
           <hr />
-          <button className="clear-filter">
+          <button className="clear-filter" onClick={onClickClearFilters}>
             <p>Clear Filters</p>
           </button>
         </div>
